pyproject.toml
[project]
name = "taskpilot"
version = "0.1.0"
description = ""
authors = [
    {name = "Aleksandr Terekhov",email = "cahdro2155@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "redis[async] (>=6.4.0,<7.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib (>=1.7.4)",
    "bcrypt (<4.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "httpx (>=0.28.1,<0.29.0)",
    "aiosmtplib (>=4.0.2,<5.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"
pytest = "^8.3.4"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
faker = "^33.3.0"

[tool.isort]
profile = "pycharm"                # Стиль форматирования
multi_line_output = 0              # Импорты в одну строку
use_parentheses = false            # Без круглых скобок
include_trailing_comma = false     # Без запятых в конце строки
force_grid_wrap = 0                # Не разбивать строки дополнительно
line_length = 120                  # Максимальная длина строки

